[X86] clean up conv code

+++ src/ppl/nn/engines/x86/impls/src/ppl/kernel/x86/fp32/conv2d/depthwise/sse/conv2d_depthwise_fp32_sse.h
@@ -39,10 +39,10 @@ public:
 private:
     struct kernel_schedule_param {
         // Preprocessed param
-        int32_t padded_ch;
+        int64_t padded_ch;
 
         // Kernel tunning
-        int32_t ow_kr_blk;
+        int64_t ow_kr_blk;
     } schedule_param_;
 
     void init_preproc_param();
